refactor: extract ovf deploy logic
- Moves OVF deployment types, orchestration, and tests from `driver_*.go` into dedicated `ovf_*.go` files for better logical organization. No behavior changes.
- Also moves content from `ovf_network*.go` into `ovf*.go` files since `buildOvfNetworkMappings` is a small helper used only by OVF import parameter creation.

Note: `ovf_archive.go` and `ovf_sanitize.go` remain unchanged since they are cross-package helpers.
